requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
av==15.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
ctranslate2==4.6.0
cycler==0.12.1
decorator==5.2.1
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
docopt==0.6.2
duckdb==1.4.1
einops==0.8.1
faster-whisper==1.2.0
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.9.23
Flask==3.1.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
hf-xet==1.1.10
httplib2==0.31.0
huggingface-hub==0.35.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.43.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.2.1
nltk==3.9.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.19.2
optuna==4.5.0
packaging==23.2
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
polars==1.35.1
polars-runtime-32==1.35.1
pooch==1.8.2
portalocker==3.2.0
primePy==1.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
silero-vad==6.0.0
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
SQLAlchemy==2.0.43
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.8.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/refs/heads/master.zip#sha256=7455f609728958b7e6801c179da18bdb25c688b52a3170e7d2aa6b5aef292003
zipp==3.23.0
